third_party/blink/web_tests/external/wpt/css/css-text/parsing/text-spacing-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['text-spacing'] = "normal" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "none" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "auto" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "no-autospace" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "trim-start" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "space-all" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "normal normal" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "normal trim-start" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "no-autospace normal" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "no-autospace space-all" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "no-autospace trim-start" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "trim-start normal " should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "normal no-autospace" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "space-all no-autospace" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-spacing'] = "trim-start no-autospace" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
